--- backend/apps/group/templates/group/mail/new_admin.html
@@ -1,8 +1,0 @@
-<small>English version bellow</small>
-<h2>Vous etes devenu admin de {{ group.name }}!</h2>
-<p>Bonjour {{ user.first_name|title }}, tu es devenu admin de {{ group.name }}.</p>
-<p>Tu peux maintenant éditer le club, ses events et ses news <a href='https://nantral-platform.fr{% url group.app|add:":update" group.slug %}'>ici</a>.</p>
-
-<h2>You're now admin of {{ group.name }}!</h2>
-<p>Hi {{ user.first_name|title }}, you are now admin of {{ group.name }}.</p>
-<p>You can now edit the club, its events and news <a href='https://nantral-platform.fr{% url group.app|add:":update" group.slug %}'>here</a>.</p>
