--- backend/apps/group/templates/group/mail/remove_admin.html
@@ -1,7 +1,0 @@
-<small>English version bellow</small>
-<h2>Tu as été retiré des admins de {{ group.name }}!</h2>
-<p>Bonjour {{ user.first_name|title }}, tu n'es plus admin de {{ group.name }}.</p>
-<p>Si tu penses que c'est une erreur contact les admins de la platforme.</p>
-
-<h2>You're no longer admin of {{ group.name }}!</h2>
-<p>If you believe this is an error please contact an admin.</p>
